+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -113,6 +113,12 @@ New check aliases
 Changes in existing checks
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+- Improved :doc:`bugprone-narrowing-conversions
+  <clang-tidy/checks/bugprone/narrowing-conversions>` check by adding the
+  ``WarnOnTimeTNarrowingConversion`` option to diagnose conversions from
+  ``time_t`` values to integer types that may not preserve the full range of
+  ``time_t``. This option is disabled by the default.
+
 - Improved :doc:`cppcoreguidelines-pro-type-member-init
   <clang-tidy/checks/cppcoreguidelines/pro-type-member-init>` check by treating
   ``std::array`` the same as built-in arrays when `IgnoreArrays` option is enabled.